requirements.txt
https://www.djangoproject.com/download/1.7c1/tarball/
markdown==2.4.1
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.7
djangorestframework==2.3.14
gunicorn==19.0.0
psycopg2==2.5.3
pytz==2014.4
six==1.7.3
static3==0.5.1
wsgiref==0.1.2
newrelic==2.22.1.20
pillow==2.5.1
